requirements.txt
Django==1.11.28
anyjson==0.3.3
redis<=2.10.6
celery==4.2.1
django-picklefield==0.3.0
kombu==4.2.0
html5lib==0.999
psycopg2-binary==2.8.6
pyparsing==2.4.2
python-dateutil>=1.5
python-openid==2.2.5
wsgiref==0.1.2
gunicorn==19.9
requests==2.21.0
unicodecsv==0.9.0
dj_database_url==0.3.0
django_webtest>=1.9
webtest==2.0.18
bleach==1.4.1
boto==2.27.0
django-ses==0.6.0
validate_email==1.2
oauth2client==1.2
django-auth-ldap==1.6.1
rollbar==0.12.1
django_celery_results==1.0.4
django_celery_beat==1.6.0
pandas==0.24.2
azure-storage-queue==2.1.0
unidecode==1.2.0
twilio==6.63.2
httplib2==0.18.0
reportlab==3.5.59
xhtml2pdf==0.2.5
django-extensions==1.9.0
django-jsonfield==1.4.1
django-braces==1.14.0
python2-secrets==1.0.5


